doc: add README
This is the information that was stored in /etc/conf.d/net while these
scripts were part of OpenRC. Moving it to the doc directory means that
there is no longer a conf.d/net file to worry about with upgrades. This
is not a regression since the default conf.d/net was only comments and
was meant to be replaced.