net/iproute2: IPv6 DAD delay must be non-fatal.
Commit 50c9fd7827fa58d27a593c8c480068f85c047cec introduced a subtle
accidental change in the handling of IPv6 tentative addresses.

In an environment where the DAD completion time exceeds the configured
timeout (which was also reduced to 5 seconds, while the docs say 10
seconds), the interface was recorded as failing to start when the
timeout was hit without DAD completion. This means that the interface
was then marked as failed and later services did not start. No rollback
took place either.

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>