net/iproute2: no automatic scope=host on loopback.
Busybox 'ip address add' does not support the 'scope' argument.

It DOES support the scope argument for 'ip address show', as well as all
of 'ip route'. The kernel does get automatic scope correct now, so drop
any automatic setting of scope to host when it was not otherwise
specified, and let the kernel make it's own decision.

X-Gentoo-Bug: 487208
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=487208
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>