net/iproute2: clarify "waiting for IPv6 addresses"
Per bug 636846, a user thought that "waiting for IPv6 addresses" was
SLAAC (Stateless Autoconfiguration). In Linux, SLAAC is entirely
kernel-side, and the waiting is actually for DAD (duplicate address
detection) on link-local IPv6 addresses.

- Improve the message to include both DAD & tentative.
- If --verbose is used, print the tentative addresses.

If either of the accept_dad sysctls are set to zero, then the kernel
should NOT mark any addresses as tentative.
- net.ipv6.conf.all.accept_dad=0
- net.ipv6.conf.$IFACE.accept_dad=0

Bug: https://bugs.gentoo.org/636846
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>