init.d/net.lo: add a configurable presence timeout
Sometimes one may want to wait for a particular interface to show up
when starting its service.
For example if a "net.foo" service is in the "default" runlevel so it
provides (or co-provides) the "net" service and it takes a while for the
"foo" interface to initialize and show up in the system during boot this
interface initialization will race with starting of this "net.foo" service
by the service manager - if the interface hasn't shown up yet the service
won't be able to start (and so will services that depend on it).

This setting specifies how long we'll wait for an interface to show up
in this case (in seconds).
For backward-compatibility the default is 0 (don't wait at all) - this
matches the existing behavior of netifrc, so existing deployments aren't
affected by this change.

This new setting is similar to an already present "wait for carrier
timeout" setting.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>