doc/net.example.Linux: document !wireless better
Users that used layer2 wireless configuration external to netifrc and
had previously configured netifrc with:
`modules="!iwconfig !wpa_supplicant"`
were surprised by the introduction of the iw & iwd modules, because the
modules tries to auto-configure the wireless interfaces.

Document that `modules="!wireless"` is a better configuration knob for
this situation.

Bug: https://bugs.gentoo.org/717226
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>