net/iproute2.sh: Add initial support for network namespaces
This adds initial support for network namespaces. An interface can be
assigned to a network namespace with `netns_${IFACE}`. The script will
move the interface to the namespace if it is not already there.

This adds a helper functions to `functions.sh` called `_netns`, this
function facilitates scripts working withing network namespaces. It
allows eching to sysfs etc files, globbing, and arbitrary commands
within a network namespace.

This uses a wrapper for the `ip` command so all calls to it will add
`-n ${netns}` so it's operating in the netns. Basic interface
configuration is tested and working.

Signed-off-by: Patrick McLean <chutzpah@gentoo.org>
Signed-off-by: Lars Wendler <polynomial-c@gentoo.org>