net/apipa.sh: Simplify and address the remaining portability issues
Re-factor the over-generalised seeding function into a _random_uint16
function. Have it use a simpler, faster method to collect entropy, with
the aid of od(1). Shorten the previously rambling comment.

Simplify the _random_apipa_octets function. Clamp the seed to the range
0-32767 for maximal portability. Convey the seed as a formal parameter,
rather then inject it. Use a simpler method to produce the octets, running
awk(1) only once. Format the random float, so that awk(1) is prevented
from using scientific notation to represent certain numbers.

Change a variable name in the subshell responsible for reading the list of
octet pairs, so as to be less confusing to future readers.

Note that the portability issue mentioned by commit 31a05f1 is addressed.
The code should now be fully POSIX-compliant, save for the continued use
of the local keyword, upon which it does not rely.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Lars Wendler <polynomial-c@gentoo.org>