net/iproute2.sh: in _get_mac_address, don't return multiple addresses
Currently if a device has virtual functions configured, ip will list the mac
addresses for the virtual functions on a device as well as the device itself.
This makes _get_mac_address return these addresses as well, causing functions
that consume this output to fail in "interesting" ways. This makes sure it only
returns 1 address.

Also don't set the mac variable to the address of the interface from sysfs, then
proceed to overwrite it with a call to "ip".

Signed-off-by: Patrick McLean <chutzpah@gentoo.org>