net/dummy.sh: preload module, create interface conditionally
The dummy module has a numdummies parameter which will cause it to
automatically create N dummy interfaces at load time.  Because creating
a dummy interface causes the kernel to load the module if it is not
already loaded, then if the name of the interface to be created matches
the name of one of the interfaces the kernel creates, an error will be
returned indicating that the interface already exists, despite it not
existing before the command was invoked.

Ensure we load the module before attempting to create any interface,
then only create the interface if it does not already exist, otherwise
simply configure it.

See: https://serverfault.com/q/839430
Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>
Closes: https://github.com/gentoo/netifrc/pull/43